dependency fix

+++ src/server/modules/user/Scope.entity.ts
@@ -16,7 +16,7 @@ export class UserScopeEntity {
 
   @Column(
     {
-      type: 'tinyint',
+      type: 'smallint',
       nullable: false,
     },
   )
